oc 10 will only operate with the marketplace - kill all appstore code
